moved slices and store to redux folder for better organization
