+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import App from './components/App';
-import store from './store/store';
+import store from './redux/store';
 
 // wrapping App in a Provider component with store prop makes the store available to the App
 const container = document.querySelector('#root') as HTMLElement;
