+++ src/redux/slices/counterSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, type PayloadAction, type ThunkAction } from '@reduxjs/toolkit';
-import { type RootState } from '../../store/store';
+import { type RootState } from '../store';
 
 interface CounterState {
   value: number
